requirements.txt
# Core Python
python>=3.9

# Telegram scraping
telethon>=1.34.0
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Data transformation
dbt-core>=1.7.0
dbt-postgres>=1.7.0

# Computer vision
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
pillow>=10.0.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Orchestration
dagster>=1.5.0
dagster-postgres>=0.20.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
